+++ Content/Python/UEFN_Toolbelt/tools/__init__.py
@@ -66,3 +66,4 @@ from . import prefab_migrator
 from . import ui_icon_import
 from . import level_health
 from . import prefab_stamp
+from . import activity_log_tools